// libs/core/models/anime-source.ts
/** Anime source. */
export enum AnimeSource {
	FourKomaManga = 'Four KomaManga',
	Book = 'Book',
	CardGame = 'Card Game',
	Game = 'Game',
	LightNovel = 'Light Novel',
	Manga = 'Manga',
	MixedMedia = 'Mixed Media',
	Music = 'Music',
	Novel = 'Novel',
	Original = 'Original',
	PictureBook = 'Picture Book',
	Radio = 'Radio',
	VisualNovel = 'Visual Novel',
	WebManga = 'Web Manga',
	WebNovel = 'Web Novel',
	Other = 'Other',
	Unknown = 'Unknown',
}